+++ domain/errors.go
@@ -290,3 +290,27 @@ type PriceNotFoundForPoolLiquidityCapError struct {
 func (e PriceNotFoundForPoolLiquidityCapError) Error() string {
 	return fmt.Sprintf("price not found (zero) for denom %s", e.Denom)
 }
+
+type FailCastCanonicalOrderbookEntryError struct {
+	BaseQuoteKey string
+}
+
+func (e FailCastCanonicalOrderbookEntryError) Error() string {
+	return fmt.Sprintf("failed to cast orderbook entry for key (%s)", e.BaseQuoteKey)
+}
+
+type FailSplitCanonicalOrderBookKeyError struct {
+	BaseQuoteKey string
+}
+
+func (e FailSplitCanonicalOrderBookKeyError) Error() string {
+	return fmt.Sprintf("failed to split base and quote denom from key %s", e.BaseQuoteKey)
+}
+
+type FailCastCanonicalOrderbookKeyError struct {
+	BaseQuoteKey string
+}
+
+func (e FailCastCanonicalOrderbookKeyError) Error() string {
+	return fmt.Sprintf("failed to cast key with value %v, expected string", e.BaseQuoteKey)
+}